requirements.txt
anyio==4.9.0
argcomplete==3.5.3
asgiref==3.9.1
azure-core==1.35.0
azure-storage-blob==12.25.1
bleach==6.2.0
build==1.2.2.post1
CacheControl==0.14.3
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
cleo==2.1.0
colorama==0.4.6
commitizen==3.31.0
comtypes==1.4.11
coverage==7.9.2
crashtest==0.4.1
cryptography==45.0.5
decli==0.6.3
distlib==0.3.9
dj-database-url==3.0.1
Django==5.0.14
django-storages==1.14.6
djangorestframework==3.16.0
dulwich==0.22.8
factory_boy==3.3.3
Faker==37.4.0
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
installer==0.7.0
isodate==0.7.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
Jinja2==3.1.6
keyring==25.6.0
Markdown==3.8.2
MarkupSafe==3.0.2
more-itertools==10.7.0
msgpack==1.1.1
mssql-django==1.5
packaging==25.0
pbs-installer==2025.7.12
pkginfo==1.12.1.2
platformdirs==4.3.8
pluggy==1.6.0
poetry==2.1.3
poetry-core==2.1.3
prompt_toolkit==3.0.51
psycopg==3.2.9
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.2
pyodbc==5.2.0
pyproject_hooks==1.2.0
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
questionary==2.1.0
RapidFuzz==3.13.0
requests==2.32.4
requests-toolbelt==1.0.0
setuptools==57.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
termcolor==2.5.0
tomlkit==0.13.3
trove-classifiers==2025.5.9.12
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.31.2
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.36.2
xattr==1.2.0
zstandard==0.23.0
